requirements.txt
aiofiles==24.1.0
alembic==1.13.1
anyio==4.8.0
bcrypt==4.1.2
bidict==0.23.1
bleach==6.2.0
blinker==1.7.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.0
cssmin==0.2.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.2
Flask-Assets==2.1.0
Flask-Bcrypt==1.0.1
Flask-CKEditor==1.0.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
get-annotations==0.1.2
googleapis-common-protos==1.66.0
greenlet==3.0.3
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.3
jsmin==3.0.1
Mako==1.3.2
Markdown==3.7
MarkupSafe==2.1.5
oauthlib==3.2.2
pillow==11.1.0
protobuf==5.29.3
psutil==6.1.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.29
typing_extensions==4.10.0
urllib3==2.3.0
webassets==2.0
webencodings==0.5.1
Werkzeug==3.0.2
wsproto==1.2.0
WTForms==3.1.2
yandex-cloud-ml-sdk==0.2.4
yandexcloud==0.330.0
gunicorn==23.0.0
eventlet==0.39.0
jedi==0.19.2
parso==0.8.4